Revert "- Adding tests for: popular topics page, product solutions page, contact support page, submitted aaq questions for freemium products, freemium aaq form (#5812)"
This reverts commit db58b90a728edf935e365207ca071690b05a6de2.
