--- playwright_tests/messages/AAQ_messages/aaq_form_page_messages.py
@@ -1,10 +1,0 @@
-class AAQFormMessages:
-    COMPLETED_MILESTONE_ONE = "Change Product"
-    COMPLETED_MILESTONE_TWO = "Explore Solutions"
-    IN_PROGRESS_MILESTONE = "Get Support"
-    SUMO_PAGE_INTRO = ("Be nice. Our volunteers are Mozilla users just like you, who take the "
-                       "time out of their day to help.")
-    INFO_CARD = ("Be descriptive. Saying “playing video on YouTube is always choppy” will help us "
-                 "understand the issue better than saying “something is wrong” or “the app is "
-                 "broken”.")
-    ERROR_MESSAGE = "This field is required."
