--- playwright_tests/messages/AAQ_messages/aaq_widget.py
@@ -1,5 +1,0 @@
-class AAQWidgetMessages:
-    FREEMIUM_PRODUCTS_AAQ_WIDGET_BUTTON_TEXT = "Ask Now"
-    PREMIUM_PRODUCTS_AAQ_WIDGET_BUTTON_TEXT = "Contact Support"
-    FREEMIUM_AAQ_SUBHEADING_TEXT_SIGNED_OUT = "Sign in/up to post your question and get help."
-    PREMIUM_AAQ_SUBHEADING_TEXT = "Send a message to our support team."
