--- playwright_tests/messages/contact_support_page_messages/contact_support_messages.py
@@ -1,23 +1,0 @@
-class ContactSupportMessages:
-    PAGE_URL = "https://support.allizom.org/en-US/questions/new"
-    PAGE_URL_CHANGE_PRODUCT_REDIRECT = PAGE_URL + "?q=change_product"
-    CURRENT_MILESTONE = "Select Product"
-    MAIN_HEADER = "Contact Support"
-    SUBHEADING = "Which product do you need help with?"
-    PRODUCT_CARDS_SUBHEADING = {
-        "Firefox": "Web browser for Windows, Mac and Linux",
-        "Firefox for Android": "Web browser for Android smartphones and tablets",
-        "Firefox for iOS": "Firefox for iPhone, iPad and iPod touch devices",
-        "Firefox for Enterprise": "Firefox Quantum for businesses",
-        "MDN Plus": "MDN Plus provides a custom experience for MDN supporters.",
-        "Firefox Reality": "Firefox for Virtual Reality headsets",
-        "Mozilla VPN": "VPN for Windows 10, Android and iOS devices",
-        "Firefox Private Network": "Browse securely on public Wi-Fi",
-        "Firefox Relay": "Service that lets you create aliases to hide your real email",
-        "Pocket": "The web’s most intriguing articles",
-        "Hubs": "Virtual 3D meeting spaces for collaborating with friends, family, "
-                "and colleagues on your browser or VR headset",
-        "Thunderbird": "Email software for Windows, Mac and Linux",
-        "Firefox Focus": "Automatic privacy browser and content blocker",
-        "Mozilla Account": "Mozilla account is the account system for Mozilla"
-    }
