--- playwright_tests/messages/forums_messages/support_forums_page_messages.py
@@ -1,4 +1,0 @@
-class SupportForumsPageMessages:
-    PAGE_URL = "https://support.allizom.org/en-US/questions/"
-    PAGE_TITLE = "Support Forums"
-    PAGE_SUBTITLE = "Browse by product"
