+++ playwright_tests/messages/homepage_messages.py
@@ -1,6 +1,5 @@
 class HomepageMessages:
-    STAGE_HOMEPAGE_URL = "https://support.allizom.org"
-    STAGE_HOMEPAGE_URL_EN_US = "https://support.allizom.org/en-US/"
+    STAGE_HOMEPAGE_URL = "https://support.allizom.org/en-US/"
     JOIN_OUR_COMMUNITY_CARD_TITLE = "Join Our Community"
     JOIN_OUR_COMMUNITY_CARD_DESCRIPTION = (
         "Grow and share your expertise with others. Answer questions and improve "
