+++ playwright_tests/messages/kb_article/kb_article_page_messages.py
@@ -1,4 +1,2 @@
 class KBArticlePageMessages:
     CREATE_NEW_KB_ARTICLE_STAGE_URL = "https://support.allizom.org/en-US/kb/new"
-    GET_COMMUNITY_SUPPORT_ARTICLE_LINK = ("https://support.allizom.org/en-US/kb/get-community"
-                                          "-support?exit_aaq=1")
