--- playwright_tests/messages/product_solutions_page_messages/product_solutions_messages.py
@@ -1,3 +1,0 @@
-class ProductSolutionsMessages:
-    PAGE_HEADER = " Solutions"
-    CURRENT_MILESTONE_TEXT = "Explore Solutions"
