+++ playwright_tests/pytest.ini
@@ -11,7 +11,3 @@ markers =
     userContributionTests: Tests belonging to the user contribution section.
     userQuestions: Tests belonging to the user questions section.
     messagingSystem: Tests belonging to the messaging system.
-    contactSupportPage: Tests belonging to the contact support page.
-    productSolutionsPage: Tests belonging to the product solutions page.
-    productTopicsPage: Tests belonging to the product topics page.
-    aaqPage: Tests belonging to the AAQ page.
