+++ playwright_tests/tests/contribute_pages_tests/test_contribute_article_page.py
@@ -228,9 +228,9 @@ class TestContributeArticlePage(TestUtilities):
                 self.logger.info("Verifying that the Home breadcrumb redirects to the Homepage")
                 check.equal(
                     self.get_page_url(),
-                    HomepageMessages.STAGE_HOMEPAGE_URL_EN_US,
+                    HomepageMessages.STAGE_HOMEPAGE_URL,
                     f"Expected to be on: "
-                    f"{HomepageMessages.STAGE_HOMEPAGE_URL_EN_US}"
+                    f"{HomepageMessages.STAGE_HOMEPAGE_URL}"
                     f"We are actual on: "
                     f"{self.get_page_url()}",
                 )
