+++ playwright_tests/tests/contribute_pages_tests/test_contribute_localization_page.py
@@ -181,8 +181,8 @@ class TestContributeLocalizationPage(TestUtilities):
                 self.logger.info("Verifying that the Home breadcrumb redirects to the Homepage")
                 check.equal(
                     self.get_page_url(),
-                    HomepageMessages.STAGE_HOMEPAGE_URL_EN_US,
-                    f"Expected to be on {HomepageMessages.STAGE_HOMEPAGE_URL_EN_US}"
+                    HomepageMessages.STAGE_HOMEPAGE_URL,
+                    f"Expected to be on {HomepageMessages.STAGE_HOMEPAGE_URL}"
                     f"We are actual on {self.get_page_url()}",
                 )
